Expose AzureAccount typings to other extensions
