+++ extensions/mssql/src/typings/refs.d.ts
@@ -6,4 +6,5 @@
 /// <reference path='../../../../src/sql/azdata.d.ts'/>
 /// <reference path='../../../../src/sql/azdata.proposed.d.ts'/>
 /// <reference path='../../../../src/vs/vscode.d.ts'/>
+/// <reference path='../../../azurecore/src/azurecore.d.ts' />
 /// <reference path='../../../big-data-cluster/src/bdc.d.ts'/>
